+++ lib/Crypto/Cipher/_mode_cbc.py
@@ -34,7 +34,7 @@ Ciphertext Block Chaining (CBC) mode.
 
 __all__ = ['CbcMode']
 
-from Crypto.Util.py3compat import _copy_bytes
+from Crypto.Util.py3compat import _copy_bytes, _is_immutable
 from Crypto.Util._raw_api import (load_pycryptodome_raw_lib, VoidPointer,
                                   create_string_buffer, get_raw_buffer,
                                   SmartPointer, c_size_t, c_uint8_ptr)
@@ -121,7 +121,7 @@ class CbcMode(object):
 
         self._next = [ self.encrypt, self.decrypt ]
 
-    def encrypt(self, plaintext):
+    def encrypt(self, plaintext, output=None):
         """Encrypt data with the key and the parameters set at initialization.
 
         A cipher object is stateful: once you have encrypted a message
@@ -148,27 +148,46 @@ class CbcMode(object):
           plaintext : bytes/bytearray/memoryview
             The piece of data to encrypt.
             Its lenght must be multiple of the cipher block size.
+        :Keywords:
+          output : bytearray/memoryview
+            The location where the ciphertext must be written to.
+            If ``None``, the ciphertext is returned.
         :Return:
-            the encrypted data, as a byte string.
-            It is as long as *plaintext*.
+          If ``output`` is ``None``, the ciphertext is returned as ``bytes``.
+          Otherwise, ``None``.
         """
 
         if self.encrypt not in self._next:
             raise TypeError("encrypt() cannot be called after decrypt()")
         self._next = [ self.encrypt ]
+        
+        if output is None:
+            ciphertext = create_string_buffer(len(plaintext))
+        else:
+            ciphertext = output
+            
+            if _is_immutable(output):
+                raise TypeError("output must be a bytearray or a writeable memoryview")
+        
+            if len(plaintext) != len(output):
+                raise ValueError("output must have the same length as the input"
+                                 "  (%d bytes)" % len(plaintext))
 
-        ciphertext = create_string_buffer(len(plaintext))
         result = raw_cbc_lib.CBC_encrypt(self._state.get(),
                                          c_uint8_ptr(plaintext),
-                                         ciphertext,
+                                         c_uint8_ptr(ciphertext),
                                          c_size_t(len(plaintext)))
         if result:
             if result == 3:
                 raise ValueError("Data must be padded to %d byte boundary in CBC mode" % self.block_size)
             raise ValueError("Error %d while encrypting in CBC mode" % result)
-        return get_raw_buffer(ciphertext)
 
-    def decrypt(self, ciphertext):
+        if output is None:
+            return get_raw_buffer(ciphertext)
+        else:
+            return None
+
+    def decrypt(self, ciphertext, output=None):
         """Decrypt data with the key and the parameters set at initialization.
 
         A cipher object is stateful: once you have decrypted a message
@@ -192,24 +211,44 @@ class CbcMode(object):
           ciphertext : bytes/bytearray/memoryview
             The piece of data to decrypt.
             Its length must be multiple of the cipher block size.
-
-        :Return: the decrypted data (byte string).
+        :Keywords:
+          output : bytearray/memoryview
+            The location where the plaintext must be written to.
+            If ``None``, the plaintext is returned.
+        :Return:
+          If ``output`` is ``None``, the plaintext is returned as ``bytes``.
+          Otherwise, ``None``.
         """
 
         if self.decrypt not in self._next:
             raise TypeError("decrypt() cannot be called after encrypt()")
         self._next = [ self.decrypt ]
+        
+        if output is None:
+            plaintext = create_string_buffer(len(ciphertext))
+        else:
+            plaintext = output
 
-        plaintext = create_string_buffer(len(ciphertext))
+            if _is_immutable(output):
+                raise TypeError("output must be a bytearray or a writeable memoryview")
+            
+            if len(ciphertext) != len(output):
+                raise ValueError("output must have the same length as the input"
+                                 "  (%d bytes)" % len(plaintext))
+
         result = raw_cbc_lib.CBC_decrypt(self._state.get(),
                                          c_uint8_ptr(ciphertext),
-                                         plaintext,
+                                         c_uint8_ptr(plaintext),
                                          c_size_t(len(ciphertext)))
         if result:
             if result == 3:
                 raise ValueError("Data must be padded to %d byte boundary in CBC mode" % self.block_size)
             raise ValueError("Error %d while decrypting in CBC mode" % result)
-        return get_raw_buffer(plaintext)
+
+        if output is None:
+            return get_raw_buffer(plaintext)
+        else:
+            return None
 
 
 def _create_cbc_cipher(factory, **kwargs):