+++ lib/Crypto/SelfTest/Cipher/test_ChaCha20_Poly1305.py
@@ -672,6 +672,55 @@ class TestVectorsWycheproof(unittest.TestCase):
             self.test_corrupt_decrypt(tv)
 
 
+class TestOutput(unittest.TestCase):
+
+    def runTest(self):
+        # Encrypt/Decrypt data and test output parameter
+
+        key = b'4' * 32
+        nonce = b'5' * 12
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+
+        pt = b'5' * 16
+        ct = cipher.encrypt(pt)
+
+        output = bytearray(16)
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        res = cipher.encrypt(pt, output=output)
+        self.assertEqual(ct, output)
+        self.assertEqual(res, None)
+        
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        res = cipher.decrypt(ct, output=output)
+        self.assertEqual(pt, output)
+        self.assertEqual(res, None)
+
+        import sys
+        if sys.version[:3] != '2.6':
+            output = memoryview(bytearray(16))
+            cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+            cipher.encrypt(pt, output=output)
+            self.assertEqual(ct, output)
+        
+            cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+            cipher.decrypt(ct, output=output)
+            self.assertEqual(pt, output)
+
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        self.assertRaises(TypeError, cipher.encrypt, pt, output=b'0'*16)
+        
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        self.assertRaises(TypeError, cipher.decrypt, ct, output=b'0'*16)
+
+        shorter_output = bytearray(7)
+        
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        self.assertRaises(ValueError, cipher.encrypt, pt, output=shorter_output)
+        
+        cipher = ChaCha20_Poly1305.new(key=key, nonce=nonce)
+        self.assertRaises(ValueError, cipher.decrypt, ct, output=shorter_output)
+
+
 def get_tests(config={}):
     wycheproof_warnings = config.get('wycheproof_warnings')
 
@@ -680,6 +729,7 @@ def get_tests(config={}):
     tests += list_test_cases(ChaCha20Poly1305FSMTests)
     tests += [TestVectorsRFC()]
     tests += [TestVectorsWycheproof(wycheproof_warnings)]
+    tests += [TestOutput()]
     return tests