+++ lib/Crypto/Util/strxor.py
@@ -47,17 +47,16 @@ _raw_strxor = load_pycryptodome_raw_lib("Crypto.Util._strxor",
 
 def strxor(term1, term2, output=None):
     """XOR two byte strings.
-
+    
     Args:
-        term1: bytes/bytearray/memoryview
-            The first term of the XOR operation.
-        term2: bytes/bytearray/memoryview
-            The second term of the XOR operation.
-            It must be as long as ``term1``.
-        output: None or bytearray/memoryview
-            If not ``None``, the location where the result is stored into.
-
-    Return:
+      term1 (bytes/bytearray/memoryview):
+        The first term of the XOR operation.
+      term2 (bytes/bytearray/memoryview):
+        The second term of the XOR operation.
+      output (bytearray/memoryview):
+        The location where the result must be written to.
+        If ``None``, the result is returned.
+    :Return:
         If ``output`` is ``None``, a new ``bytes`` string with the result.
         Otherwise ``None``.
     """
@@ -93,11 +92,11 @@ def strxor_c(term, c, output=None):
     """XOR a byte string with a repeated sequence of characters.
 
     Args:
-        term : bytes/bytearray/memoryview
+        term(bytes/bytearray/memoryview):
             The first term of the XOR operation.
-        c : byte
+        c (bytes):
             The byte that makes up the second term of the XOR operation.
-        output: None or bytearray/memoryview
+        output (None or bytearray/memoryview):
             If not ``None``, the location where the result is stored into.
 
     Return: